feat(13-03): create team defense data processing module
- process_defense_data(): transforms team stats to ML features
- Computes points_allowed from schedule home/away scores
- Adds rolling 3-week features: sacks, interceptions, fumbles, TDs
- get_defense_feature_columns() and get_defense_target_columns() helpers
- Combines def_tds + special_teams_tds into total_def_tds

+++ packages/backend/src/lineupiq/models/training.py
@@ -213,7 +213,7 @@ def tune_hyperparameters(
             params = get_xgb_params(trial)
         _, scores = train_model(X, y, params=params, n_splits=n_splits, model_type=model_type)
         # Return mean negative RMSE (minimize this)
-        return -scores.mean()
+        return float(-scores.mean())
 
     # Create study and optimize
     study = create_study(direction="minimize")
